Version as of Jul 13 with bandwidth limitation

+++ WebCritterer/src/com/webcritterer/Critterer.java
@@ -5,9 +5,8 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Set;
 
-public class Critterer
-{
-    private static final int MaximumPagesToGoTo = 1000000000;
+public class Critterer {
+    private static final long MaximumPagesToGoTo = 1000000000;
     private Set<String> pagesAlreadyHit = new HashSet<String>();
     private List<String> pagesNeededToGoTo = new LinkedList<String>();
 
@@ -22,19 +21,14 @@ public class Critterer
     }
 
 
-    public void load(String url)
-    {
-        while(this.pagesAlreadyHit.size() < MaximumPagesToGoTo)
-        {
+    public void load(String url) {
+        while (this.pagesAlreadyHit.size() < MaximumPagesToGoTo) {
             String currentUrl;
             CrittererEat scramble = new CrittererEat();
-            if(this.pagesNeededToGoTo.isEmpty())
-            {
+            if (this.pagesNeededToGoTo.isEmpty()) {
                 currentUrl = url;
                 this.pagesAlreadyHit.add(url);
-            }
-            else
-            {
+            } else {
                 currentUrl = this.nextUrl();
             }
             scramble.critter(currentUrl);
